+++ doc/conf.py
@@ -22,7 +22,7 @@ sys.path.insert(0, PACKAGEDIR)
 
 # -- Project information -----------------------------------------------------
 
-project = 'scikit-gym'
+project = 'keras-gym'
 copyright = '2018, Kristian Holsheimer'
 author = 'Kristian Holsheimer'
 
@@ -136,7 +136,7 @@ html_static_path = ['_static']
 # -- Options for HTMLHelp output ---------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'scikit-gymdoc'
+htmlhelp_basename = 'keras-gymdoc'
 
 
 # -- Options for LaTeX output ------------------------------------------------
@@ -163,7 +163,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'scikit-gym.tex', 'scikit-gym Documentation',
+    (master_doc, 'keras-gym.tex', 'keras-gym Documentation',
      'Kristian Holsheimer', 'manual'),
 ]
 
@@ -173,7 +173,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'scikit-gym', 'scikit-gym Documentation',
+    (master_doc, 'keras-gym', 'keras-gym Documentation',
      [author], 1)
 ]
 
@@ -184,8 +184,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'scikit-gym', 'scikit-gym Documentation',
-     author, 'scikit-gym', 'One line description of project.',
+    (master_doc, 'keras-gym', 'keras-gym Documentation',
+     author, 'keras-gym', 'One line description of project.',
      'Miscellaneous'),
 ]
 
@@ -214,7 +214,7 @@ epub_exclude_files = ['search.html']
 
 # Example configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {
-    'python': ('https://docs.python.org/3', (None,)),
+    'python': ('https://docs.python.org/3/', (None,)),
     'numpy': ('https://docs.scipy.org/doc/numpy/', (None,)),
     'sklearn': ('https://scikit-learn.org/stable/', (None,)),
 }