+++ keras_gym/value_functions/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from .generic import GenericV, GenericQ
+from .predefined import LinearV, LinearQ