+++ skgym/value_functions/__init__.py
@@ -1,3 +1,0 @@
-# flake8: noqa
-from .generic import GenericV, GenericQ, GenericQTypeI, GenericQTypeII
-from .predefined import LinearV, LinearQ